WIP Restructuration Not final, will have more progress later
