+++ scripts/components/Home/CommunityServers.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import nitro from 'url:../../assets/discord/nitro.png'
 import PropTypes from 'prop-types'
-import Guilds from './material/guilds'
-import { shuffleArray } from '../utils'
+import Guilds from './Guilds'
+import { shuffleArray } from '../../utils'
 import Box from '@material-ui/core/Box'
 import Typography from '@material-ui/core/Typography'
 import Link from '@material-ui/core/Link'
