+++ scripts/components/Home/OfficialServers.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import Guilds from './material/guilds'
+import Guilds from './Guilds'
 import Typography from '@material-ui/core/Typography'
 import makeStyles from '@material-ui/core/styles/makeStyles'
 import Link from '@material-ui/core/Link'
